# Fix performance of Containers.rowtable