Closes #3404.

Julia 1.6

```julia
julia> using JuMP.Containers

julia> Containers.@container(x[i=1:100, j=1:50], i + j);

julia> function foo(x, header)
           names = tuple(header...)
           return [NamedTuple{names}((args..., x[i])) for (i, args) in Containers._rows(x)]
       end
foo (generic function with 1 method)

julia> function bar(x, header)
           elements = [(args..., x[i]) for (i, args) in Containers._rows(x)]
           return NamedTuple{tuple(header...)}.(elements)
       end
bar (generic function with 1 method)

julia> @assert foo(x, [:i, :j, :k]) == bar(x, [:i, :j, :k])

julia> @time foo(x, [:i, :j, :k]);
  0.006314 seconds (30.02 k allocations: 1.597 MiB)

julia> @time bar(x, [:i, :j, :k]);
  0.000154 seconds (23 allocations: 502.297 KiB)
```

Julia 1.9

```Julia
julia> using JuMP.Containers

julia> Containers.@container(x[i=1:100, j=1:50], i + j);

julia> function foo(x, header)
           names = tuple(header...)
           return [NamedTuple{names}((args..., x[i])) for (i, args) in Containers._rows(x)]
       end
foo (generic function with 1 method)

julia> function bar(x, header)
           elements = [(args..., x[i]) for (i, args) in Containers._rows(x)]
           return NamedTuple{tuple(header...)}.(elements)
       end
bar (generic function with 1 method)

julia> @assert foo(x, [:i, :j, :k]) == bar(x, [:i, :j, :k])

julia> @time foo(x, [:i, :j, :k]);
  0.378728 seconds (140.02 k allocations: 5.545 MiB)

julia> @time bar(x, [:i, :j, :k]);
  0.000079 seconds (18 allocations: 482.516 KiB)
```

I don't understand the reason for the regression in 1.9. But this is better regardless.