@odow 2023-06-07
Just to close the loop on the original benchmark:

Previously

```julia
julia> @btime JuMP.Containers.rowtable($value, $x, header=$header);
  371.580 ms (210016 allocations: 7.07 MiB)

julia> @btime test($value, $x, header=$header);
  6.915 ms (105016 allocations: 3.03 MiB)

julia> @btime JuMP.Containers.rowtable($value, $y, header=$header);
  187.218 ms (105010 allocations: 3.41 MiB)

julia> @btime test($value, $y, header=$header);
  3.433 ms (52510 allocations: 1.39 MiB)
```

This PR

```Julia
julia> @btime JuMP.Containers.rowtable(value, $x, header=$header);
  2.349 ms (55025 allocations: 1.46 MiB)

julia> @btime test(value, $x, header=$header);
  6.906 ms (105016 allocations: 3.03 MiB)

julia> @btime JuMP.Containers.rowtable(value, $y, header=$header);
  1.211 ms (27518 allocations: 625.70 KiB)

julia> @btime test(value, $y, header=$header);
  3.354 ms (52510 allocations: 1.39 MiB)
```